pyproject.toml
[tool.poetry]
name = "dialogue-summarization"
version = "0.1.0"
description = ""
authors = ["marccgrau <marcgrau@outlook.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
pre-commit = "^3.7.0"
langchain-openai = "^0.1.8"
tiktoken = "^0.7.0"
chromadb = "^0.5.0"
langchain = "^0.2.3"
langchainhub = "^0.1.18"
requests = "^2.32.3"
pandas = "^2.2.2"
scikit-learn = "^1.5.0"
tqdm = "^4.66.4"
datasets = "^2.19.2"
huggingface-hub = "^0.23.3"
openai = "^1.33.0"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.4"
jsonschema = "^4.22.0"
langchain-google-genai = "^1.0.6"
loguru = "^0.7.2"
pydantic-settings = "^2.3.3"
langchain-google-vertexai = "^1.0.5"
langchain-anthropic = "^0.1.15"
streamlit = "^1.36.0"
duckdb = "^1.0.0"
watchdog = "^4.0.1"
langchain-nvidia-ai-endpoints = "^0.1.2"
dspy-ai = "^2.4.10"
deepl = "^1.18.0"
langchain-together = "^0.1.3"
poethepoet = "^0.27.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 160
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F", "N", "D", "S"]
ignore = ["D100", "S311"] # Ignore public module docstring, Ignore standard pseudo-random generator error

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
